+++ service/watcher/chartstatus/chartstatus.go
@@ -5,9 +5,11 @@ import (
 	"fmt"
 
 	"github.com/giantswarm/apiextensions/v3/pkg/apis/application/v1alpha1"
+	"github.com/giantswarm/apiextensions/v3/pkg/clientset/versioned"
 	"github.com/giantswarm/app/v4/pkg/key"
 	"github.com/giantswarm/k8sclient/v5/pkg/k8sclient"
 	"github.com/giantswarm/k8smetadata/pkg/annotation"
+	"github.com/giantswarm/kubeconfig/v4"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 	"github.com/google/go-cmp/cmp"
@@ -15,7 +17,10 @@ import (
 	"k8s.io/apimachinery/pkg/watch"
 )
 
+const chartOperatorAppName = "chart-operator"
+
 type ChartStatusWatcherConfig struct {
+	G8sClient versioned.Interface
 	K8sClient k8sclient.Interface
 	Logger    micrologger.Logger
 
@@ -25,8 +30,10 @@ type ChartStatusWatcherConfig struct {
 }
 
 type ChartStatusWatcher struct {
-	k8sClient k8sclient.Interface
-	logger    micrologger.Logger
+	g8sClient  versioned.Interface
+	k8sClient  k8sclient.Interface
+	kubeConfig kubeconfig.Interface
+	logger     micrologger.Logger
 
 	appNamespace   string
 	chartNamespace string
@@ -34,6 +41,9 @@ type ChartStatusWatcher struct {
 }
 
 func NewChartStatusWatcher(config ChartStatusWatcherConfig) (*ChartStatusWatcher, error) {
+	if config.G8sClient == nil {
+		return nil, microerror.Maskf(invalidConfigError, "%T.G8sClient must not be empty", config)
+	}
 	if config.K8sClient == nil {
 		return nil, microerror.Maskf(invalidConfigError, "%T.K8sClient must not be empty", config)
 	}
@@ -48,9 +58,25 @@ func NewChartStatusWatcher(config ChartStatusWatcherConfig) (*ChartStatusWatcher
 		return nil, microerror.Maskf(invalidConfigError, "%T.PodNamespace must not be empty", config)
 	}
 
+	var kubeConfig kubeconfig.Interface
+	var err error
+	{
+		c := kubeconfig.Config{
+			K8sClient: config.K8sClient.K8sClient(),
+			Logger:    config.Logger,
+		}
+
+		kubeConfig, err = kubeconfig.New(c)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
+
 	c := &ChartStatusWatcher{
-		k8sClient: config.K8sClient,
-		logger:    config.Logger,
+		g8sClient:  config.G8sClient,
+		k8sClient:  config.K8sClient,
+		kubeConfig: kubeConfig,
+		logger:     config.Logger,
 
 		// We get a kubeconfig for the cluster from the chart-operator app CR
 		// which is in the same namespace as this instance of app-operator.