+++ integration/templates/chart_operator_values.go
@@ -6,12 +6,11 @@ package templates
 const ChartOperatorValues = `clusterDNSIP: 10.96.0.10
 e2e: true
 
-Installation:
-  V1:
-    Helm:
-      HTTP:
-        ClientTimeout: "30s"
-      Kubernetes:
-        WaitTimeout: "180s"
-    Registry:
-      Domain: "quay.io"`
+helm:
+  http:
+    clientTimeout: "30s"
+  kubernetes:
+    waitTimeout: "180s"
+
+registry:
+  domain: "quay.io"`