+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ### Changed
 
+- Get `Chart` CRD from the GitHub resources.
+
+### Changed
+
 - Get metadata constants from k8smetadata library not apiextensions.
 
 ### Fixed