+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ### Changed
 
+- Prepare helm values to configuration management.
 - Use `Catalog` CRs in `App` controller.
 - Reconcile to `Catalog` CRs instead of `AppCatalog`.
 - Get `Chart` CRD from the GitHub resources.
@@ -17,7 +18,7 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 ### Fixed
 
 - For the chart CR watcher get the kubeconfig secret from the chart-operator app
-CR to avoid hardcoding it. 
+CR to avoid hardcoding it.
 
 ## [4.4.0] - 2021-05-03
 
@@ -34,7 +35,7 @@ CR to avoid hardcoding it.
 
 ## [4.3.1] - 2021-03-30
 
-### Fixed 
+### Fixed
 
 - Restore chart-operator when it had been deleted.